+++ packages/authentication/modules/access/useLogin/index.ts
@@ -21,7 +21,6 @@ import type {
   LoginChangeExpiredPasswordRedirectResponse,
   LoginJWTResponse,
   LoginMfaRequest,
-  LoginRequest,
 } from '../../../types/auth'
 import { User } from '../../../types/user'
 import {
@@ -30,6 +29,7 @@ import {
 } from '../../../utils/login'
 import { CODE_VALIDATION_INITIAL_VALUES, CODE_VALIDATION_SCHEMA } from '../../mfa/constants'
 import { useCurrentProfile } from '../../profile'
+import { setProfileExpoStorage } from '../../profile/utils'
 import { DEFAULT_INITIAL_VALUES, DEFAULT_VALIDATION_SCHEMA } from './constants'
 import type { ApiClass, LoginParams, UseLoginOptions } from './types'
 
@@ -55,23 +55,28 @@ const useLogin = <TApiClass extends ApiClass = typeof AuthApi>({
       return
     }
 
-    // TODO: adapt this flow to work with mobile
-    if (!isMobilePlatform()) {
-      const user = decodeJWT<User>(response.access)
-      if (user) {
-        // TODO: handle the absolute image path on the backend
-        const baseUrl = process.env.NEXT_PUBLIC_API_BASE_URL?.replace('/v1', '')
-        let absoluteImagePath = null
-        if (user?.profile?.image) {
-          absoluteImagePath = user.profile.image.startsWith('http')
-            ? user.profile.image
-            : `${baseUrl}${user.profile.image}`
-        }
-        setCurrentProfile({
-          ...user.profile,
-          image: absoluteImagePath,
-        })
+    const isWebPlatform = !isMobilePlatform()
+
+    const user = decodeJWT<User>(response.access)
+    if (user) {
+      const API_BASE_URL = isWebPlatform
+        ? process.env.NEXT_PUBLIC_API_BASE_URL
+        : process.env.EXPO_PUBLIC_API_BASE_URL
+      const baseUrl = API_BASE_URL?.replace('/v1', '')
+      let absoluteImagePath = null
+      if (user?.profile?.image) {
+        absoluteImagePath = user.profile.image.startsWith('http')
+          ? user.profile.image
+          : `${baseUrl}${user.profile.image}`
       }
+
+      const currentProfile = {
+        ...user.profile,
+        image: absoluteImagePath,
+      }
+
+      setCurrentProfile(currentProfile)
+      await setProfileExpoStorage(currentProfile)
     }
 
     await setTokenAsync(accessKeyName, response.access, {
@@ -82,15 +87,15 @@ const useLogin = <TApiClass extends ApiClass = typeof AuthApi>({
     })
   }
 
-  const form = useForm({
+  const form = useForm<LoginParams<TApiClass>>({
     defaultValues: DEFAULT_INITIAL_VALUES as LoginParams<TApiClass>,
     resolver: zodResolver(DEFAULT_VALIDATION_SCHEMA),
-    mode: 'onBlur',
+    mode: 'onChange',
     ...loginFormOptions,
   })
 
   const mutation = useMutation({
-    mutationFn: (data: LoginRequest) => ApiClass.login(data),
+    mutationFn: (data: LoginParams<TApiClass>) => ApiClass.login(data),
     ...loginOptions, // needs to be placed bellow all overridable options
     onError: (err, variables, context) => {
       loginOptions?.onError?.(err, variables, context)
@@ -124,9 +129,10 @@ const useLogin = <TApiClass extends ApiClass = typeof AuthApi>({
         setFormApiErrors(form, err)
       }
     },
-    onSuccess: (response, variables, context) => {
-      // @ts-ignore BA-1206: fix typing
-      handleLoginSuccess(response)
+    onSuccess: async (response, variables, context) => {
+      if (!isLoginMfaResponse(response)) {
+        await handleLoginSuccess(response)
+      }
       mfaOptions?.onSuccess?.(response, variables, context)
     },
   })
@@ -134,29 +140,25 @@ const useLogin = <TApiClass extends ApiClass = typeof AuthApi>({
   return {
     form: {
       ...form,
-      // TODO: refactor types
       handleSubmit: form.handleSubmit(async (values) => {
         try {
-          await mutation.mutateAsync(values as LoginRequest)
+          await mutation.mutateAsync(values)
         } catch (error) {
           // mutateAsync will raise an error if there's an API error
         }
-        // TODO: refactor types
-      }) as any,
+      }),
     },
     mutation,
     mfaForm: {
       ...mfaForm,
-      // TODO: refactor types
-      handleSubmit: mfaForm.handleSubmit(async (values: any) => {
+      handleSubmit: mfaForm.handleSubmit(async (values) => {
         try {
-          const newValues = { ...values, ephemeralToken: mfaEphemeralToken }
+          const newValues = { token: values.code, ephemeralToken: mfaEphemeralToken || '' }
           await mfaMutation.mutateAsync(newValues)
         } catch (error) {
           // mutateAsync will raise an error if there's an API error
         }
-        // TODO: refactor types
-      }) as any,
+      }),
     },
     mfaMutation,
   }