+++ packages/components/modules/content-feed/web/PostFooter/types.ts
@@ -1,0 +1,5 @@
+import { ContentPost_post$data } from '../../../../__generated__/ContentPost_post.graphql'
+
+export interface PostFooterProps {
+  post: ContentPost_post$data
+}