+++ packages/components/modules/content-feed/web/PostList/styled.tsx
@@ -1,0 +1,8 @@
+import { Stack, styled } from '@mui/material'
+
+export const PostsListContainer = styled(Stack)(({ theme }) => ({
+  '& > :not(:last-of-type)': {
+    borderBottom: '1px solid',
+    borderColor: theme.palette.divider,
+  },
+}))