+++ packages/components/modules/navigations/web/NavMini/index.tsx
@@ -17,8 +17,6 @@ import { NavMiniProps } from './types'
 
 const NavMini: FC<NavMiniProps> = ({
   navData,
-  settings,
-  setSettings,
   LogoIcon,
   LogoProps,
   openNav,
@@ -33,6 +31,7 @@ const NavMini: FC<NavMiniProps> = ({
 
   return (
     <Box
+      data-testid="nav-mini"
       sx={{
         flexShrink: { lg: 0 },
         width: { lg: NAV_WIDTH.MINI },
@@ -41,8 +40,6 @@ const NavMini: FC<NavMiniProps> = ({
     >
       {!hideToggleButton && (
         <NavToggleButton
-          settings={settings}
-          setSettings={setSettings}
           sx={{
             top: 22,
             left: NAV_WIDTH.MINI - 12,