+++ packages/design-system/components/web/dropzones/Dropzone/types.ts
@@ -1,5 +1,6 @@
-import { CSSProperties, HTMLAttributes, InputHTMLAttributes } from 'react'
+import { CSSProperties, FC, HTMLAttributes, InputHTMLAttributes, ReactElement } from 'react'
 
+import { SvgIconProps } from '@mui/material'
 import { Theme } from '@mui/material/styles'
 import type { Accept, DropzoneOptions } from 'react-dropzone'
 
@@ -22,11 +23,17 @@ export interface BaseDropzoneProps {
   maxFileSize?: number
   DropzoneOptions?: Partial<DropzoneOptions>
   InputProps?: InputHTMLAttributes<HTMLInputElement>
-  title?: string | JSX.Element
+  title?: string | ReactElement
   includeActionButton: boolean
   InputContainerStyle?: CSSProperties
   multiple?: boolean
   onFileClick?: (selectedFile: string | File | Blob, index?: number) => void
+  DropzonePreview?: FC<DropzonePreviewProps>
+  DropzonePreviewProps?: Partial<DropzonePreviewProps>
+  DropZoneIcon?: FC<SvgIconProps>
+  DropZoneIconProps?: Partial<SvgIconProps>
+  DropZoneCancelIcon?: FC<SvgIconProps>
+  DropZoneCancelIconProps?: Partial<SvgIconProps>
 }
 
 export interface DropzoneAsSingleBase64Props extends BaseDropzoneProps {