+++ packages/utils/hooks/useCookie/constants.ts
@@ -1,0 +1,2 @@
+export const MISSING_COOKIE_STORE_ERROR =
+  'Cookie store has not been initialized. Make sure CookieProvider is used in the application.'